test: split runtime route selection coverage
